+++ bloodDonors/src/main/java/com/training/ifaces/Dao.java
@@ -5,9 +5,9 @@ public interface Dao<T> {
 
 	public int add(T t) throws SQLException;
    public List<T> findAll() throws SQLException; 
-   public T findById(int id) throws SQLException;
+   public T findById(long id) throws SQLException;
    public int update(T t) throws SQLException;
-   public int remove(int id) throws SQLException;
+   public int remove(long id) throws SQLException;
    
    
 }
